* tree-vectorizer.c (vect_can_force_dr_alignment_p): 	Temporarily cast PREFERRED_STACK_BOUNDARY to unsigned, until 	the next merge.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lno-branch@86564 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-vectorizer.c
@@ -1440,7 +1440,7 @@ vect_can_force_dr_alignment_p (tree decl, unsigned int alignment)
        PREFERRED_STACK_BOUNDARY is honored by all translation units.
        However, until someone implements forced stack alignment, SSE
        isn't really usable without this.  */   
-    return (alignment <= PREFERRED_STACK_BOUNDARY);
+    return (alignment <= (unsigned int) PREFERRED_STACK_BOUNDARY);
 }
